Remove secret from notebookControllerState, fix modal close bug
